pyproject.toml
[project]
name = "zerion-mcp-server"
version = "0.1.0"
description = "MCP Server for Zerion API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.8.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "uvicorn>=0.23.0",
    "starlette>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",  # HTTP mocking for httpx
]

[project.scripts]
zerion-mcp-server = "zerion_mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--cov=zerion_mcp_server",
    "--cov-report=term-missing",
    "--cov-report=html",
]
